+++ test/index.js
@@ -12,13 +12,16 @@ test('exports an "updateNodejsNotifier" function', (t) => {
   t.is(typeof lib.updateNodejsNotifier, 'function')
 })
 
+// assumes current Node.js version is >=4
 test('updateNodejsNotifier({ versions }), no matching alert', (t) => {
   const versions = [
     { version: 'v3.4.5', date: '', files: [], modules: '', lts: false },
     { version: 'v2.3.4', date: '', files: [], modules: '', lts: false },
     { version: 'v1.2.3', date: '', files: [], modules: '', lts: false }
   ]
-  t.is(lib.updateNodejsNotifier({ versions }), false)
+  const notify = () => { notify.called = true }
+  t.is(lib.updateNodejsNotifier({ notify, versions }), false)
+  t.falsy(notify.called)
 })
 
 test('updateNodejsNotifier({ versions }), matching alert', (t) => {
@@ -27,9 +30,16 @@ test('updateNodejsNotifier({ versions }), matching alert', (t) => {
     { version: 'v2.3.4', date: '', files: [], modules: '', lts: false },
     { version: 'v1.2.3', date: '', files: [], modules: '', lts: false }
   ]
+  const notify = (options) => {
+    const message = options.message || ''
+    t.truthy(message)
+    notify.called = true
+  }
   t.is(lib.updateNodejsNotifier({
     current: 'v1.2.3',
+    notify,
     stableMajor: true,
     versions
   }), true)
+  t.truthy(notify.called)
 })