+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 -   if caller passes "versions" in, don't fetch versions (better for testing)
 
+-   allow caller to pass in "notify" function (better for testing)
+
 
 ## 1.1.0 - 2016-11-01